+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -263,7 +263,7 @@ class GeckoEngineSession(
                 // Log the error. There is nothing we can do otherwise.
                 logger.error("Save to PDF failed.", throwable)
                 notifyObservers {
-                    onSaveToPdfError(throwable)
+                    onSaveToPdfException(throwable)
                 }
                 GeckoResult()
             },